retriever/requirements.txt
openai>=1.0.0
opensearch-py>=2.4.0
psycopg2-binary>=2.9.0
requests>=2.31.0
minio>=7.0.0
